Fix bug where the wrong chapter would download
Previously, the download_manga() function would assume that the list of
chapters in a manga would have a one-to-one correspondence with their
index in the chapter list. This is not the case if a manga has chapters
with decimals or a chapter 0 or some other nonstandard numbering. The
get_chapter_urls() code has been rewritten to use a dictionary instead of
a list so that the download_manga() function no longer relies on any
one-to-one correspondence between chapter number and index in the list
of chapters.