Add support for "manga_name" alongside "Manga Name"
Mangas can be inputted in the url format (e.g, nana_to_kaoru) as well as
the format displayed on the page (e.g, Nana to Kaoru). Fix for bug where
the script would die with an IndexError when the manga name was inputted
in the url format because the manga name would not be filtered out of
the link text (the key/index for the chapter_urls dictionary) for each
chapter link.